[BUGFIX] Fix disabled Schedule button
