--- common/containers/Tabs/ScheduleTransaction/components/Fields/WindowSize/WindowSizeFieldFactory/index.ts
@@ -1,1 +1,0 @@
-export * from './WindowSizeFieldFactory';
