+++ common/containers/Tabs/ScheduleTransaction/components/index.ts
@@ -1,6 +1,6 @@
 export * from './Fields/WindowStart/WindowStartField';
 export * from './Fields/TimeBounty/TimeBountyField';
-export * from './Fields/WindowSize/WindowSizeField';
+export * from './Fields/WindowSizeField';
 export * from './Fields/SchedulingToggle/SchedulingToggle';
 export * from './Fields/ScheduleGasPriceField';
 export * from './Fields/ScheduleGasLimitField';
